requirements.txt
openai>=1.0.0
duckduckgo-search>=5.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pyautogui>=0.9.54
pillow>=10.0.0
pyperclip>=1.8.2
requests>=2.31.0
beautifulsoup4>=4.12.0
